requirements.txt
pyTelegramBotAPI==4.3.1
bs4==0.0.1
requests==2.25.1
pytz==2021.1
prettytable==3.0.0
aiohttp==3.8.3
python-dotenv==1.0.0